backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
sentence-transformers==3.1.1
faiss-cpu==1.9.0
ollama==0.3.3
pydantic==2.9.2